// src/request-type.ts
export interface InitializeParams extends vscode.InitializeParams {
  capabilities: ClientCapabilities;
}

export interface ClientCapabilities extends vscode.ClientCapabilities {
  /**
   * The client provides support for workspace/xfiles.
   */
  xfilesProvider?: boolean;

  /**
   * The client provides support for textDocument/xcontent.
   */
  xcontentProvider?: boolean;

  /**
   * The client provides support for cache/get and cache/set methods
   */
  xcacheProvider?: boolean;

  /**
   * The client supports receiving the result solely through $/partialResult
   * notifications for requests from the client
   * to the server.
   */
  streaming?: boolean;
}

export interface ServerCapabilities extends vscode.ServerCapabilities {
  xworkspaceReferencesProvider?: boolean;
  xdefinitionProvider?: boolean;
  xdependenciesProvider?: boolean;
  xpackagesProvider?: boolean;

  /**
   * The server supports receiving results solely through $/partialResult
   * notifications for requests from the server to
   * the client.
   */
  streaming?: boolean;
}

export interface InitializeResult extends vscode.InitializeResult {
  capabilities: ServerCapabilities;
}